+++ examples/helloworld/build.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env
 
-browserify -t coffeeify --extension=".coffee" -o bundle.js index.coffee
+browserify -t babelify -o bundle.js index.js