+++ examples/website-router/build.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env
-
-browserify -t babelify -o bundle.js index.js